+++ src/Phan/Language/Type/FloatType.php
@@ -61,6 +61,8 @@ class FloatType extends ScalarType
         if ($other instanceof ScalarType) {
             return $other instanceof FloatType || (!$context->isStrictTypes() && parent::canCastToDeclaredType($code_base, $context, $other));
         }
-        return $other instanceof CallableType || $other instanceof MixedType;
+        return $other instanceof CallableType ||
+            $other instanceof TemplateType ||
+            $other instanceof MixedType;
     }
 }