+++ src/Phan/Language/Type/StringType.php
@@ -49,7 +49,9 @@ class StringType extends ScalarType
         if ($other instanceof ScalarType) {
             return $other instanceof StringType || (!$context->isStrictTypes() && parent::canCastToDeclaredType($code_base, $context, $other));
         }
-        return $other instanceof CallableType || $other instanceof MixedType;
+        return $other instanceof CallableType ||
+            $other instanceof TemplateType ||
+            $other instanceof MixedType;
     }
 
     /**