+++ src/Phan/Language/Type/TrueType.php
@@ -93,6 +93,7 @@ final class TrueType extends ScalarType
     {
         return $other->isInBoolFamily() ||
             $other instanceof MixedType ||
+            $other instanceof TemplateType ||
             (!$context->isStrictTypes() && parent::canCastToDeclaredType($code_base, $context, $other));
     }