+++ tests/files/src/0226_internal_tostring_parameter_undeclared.php
@@ -2,7 +2,7 @@
 
 /**
  * @return urlstring
- * @suppress PhanTypeMismatchReturn
+ * @suppress PhanTypeMismatchReturnProbablyReal
  */
 function foo() {return 'a';}