+++ NEWS.md
@@ -15,6 +15,7 @@ New features (Analysis):
   This is useful when needing to preserve method signature compatibility in a method override, or when a parameter will become mandatory in a future backwards incompatible release of a project.
 + Emit `PhanTypeMismatchArgumentProbablyReal` instead of `PhanTypeMismatchArgument` when the inferred real type of an argument has nothing in common with the phpdoc type of a user-defined function/method.
   This is usually a stronger indicator that the phpdoc parameter type is inaccurate/incomplete or the argument is incorrect.
++ Emit `PhanTypeMismatchReturnProbablyReal` instead of `PhanTypeMismatchReturn` when the inferred real return type has nothing in common with the declared phpdoc return type of a user-defined function/method. (#4028)
 + Emit `PhanTypeMismatchArgumentInternalProbablyReal` instead of `PhanTypeMismatchArgumentInternal` in a few more cases.
 + Be stricter about checking if callables/closures have anything in common with other types.
 + Preserve more specific phpdoc types when the php 8.0 `mixed` type is part of the real type set.